+++ shotmanager/prefs/features.py
@@ -110,9 +110,9 @@ def draw_features_prefs(mode, layout):
     icon = config.icons_col["ShotManager_CamGPVisible_32"]
     subrow.prop(props, "display_storyboard_in_properties", text="", icon_value=icon.icon_id)
     subSubrow = subrow.row()
-    subSubrow.enabled = props != prefs
+    # subSubrow.enabled = props != prefs
     subSubrow.scale_x = 0.9
-    subSubrow.operator("uas_shot_manager.greasepencil_template_panel", text="", icon="LONGDISPLAY")
+    subSubrow.operator("uas_shot_manager.greasepencil_template_panel", text="", icon="LONGDISPLAY").mode = mode
     subrow.label(text="Storyboard")
 
     ################