+++ shotmanager/config/config.py
@@ -38,7 +38,7 @@ def initGlobalVariables():
         devDebug = False
 
     # change this value to force debug at start time
-    devDebug = True
+    devDebug = False
 
     global devDebug_lastRedrawTime
     devDebug_lastRedrawTime = -1
@@ -85,6 +85,7 @@ def initGlobalVariables():
     # keymaps ##########
     global gAddonKeymaps
     gAddonKeymaps = []
+
 
 def releaseGlobalVariables():