+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.0.21 (2022-05-10)
+
+- Display the shots names over the storyboard frames when the camera is hidden
+
+
 # 2.0.20 (2022-05-06)
 
 - Added key mapping for previous and next shots, previous and next grease pencil key frame